Also scan workbenches for controllers with menu's
The current implementation only checks the app's core controllers, but its generally cleaner to construct your application in separate modules to allow re-use of modules and easier upgrade/management. 

The vendor folder should probably also be checked, but that could make the generation slower. The path should really not be hardcoded and an configurable array with paths would be cleaner (to also allow for inclusion of non-app and non-workbench entries)